+++ hera_cal/__init__.py
@@ -4,9 +4,9 @@ import firstcal
 import redcal
 import cal_formats
 import version
-import abscal
 import io
 import delay_filter
+import abscal
 import lstbin
 import smooth_cal
 import apply_cal
