+++ hera_cal/tests/test_abscal.py
@@ -217,6 +217,21 @@ class Test_AbsCal_Funcs:
         relevant_mfiles = hc.abscal.match_times(dfiles[0], mfiles)
         nt.assert_equal(len(relevant_mfiles), 0)
 
+    def test_rephase_vis(self):
+        dfile = os.path.join(DATA_PATH, 'zen.2458043.12552.xx.HH.uvORA')
+        mfiles = map(lambda f: os.path.join(DATA_PATH, f), ['zen.2458042.12552.xx.HH.uvXA'])
+        m, mf, mantp, mant, mfr, mt, ml, mp = io.load_vis(mfiles)
+        d, df, dantp, dant, dfr, dt, dl, dp = io.load_vis(dfile)
+        bls = odict(map(lambda k: (k, dantp[k[0]] - dantp[k[1]]), d.keys()))
+
+        # basic execution
+        new_m, new_f = hc.abscal.rephase_vis(m, ml, dl, bls, dfr)
+
+        k = new_m.keys()[0]
+        nt.assert_equal(new_m[k].shape, d[k].shape)
+        nt.assert_true(new_f[k][0].min())
+        nt.assert_false(new_f[k][-1].max())
+
 
 class Test_AbsCal:
